requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
av==14.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
distro==1.9.0
docstring_parser==0.16
eval_type_backport==0.2.2
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
livekit==1.0.7
livekit-agents==1.0.20
livekit-api==1.0.2
livekit-plugins-cartesia==1.0.20
livekit-plugins-noise-cancellation==0.2.1
livekit-plugins-openai==1.0.20
livekit-plugins-rime==1.0.20
livekit-plugins-silero==1.0.11
livekit-plugins-turn-detector==1.0.11
livekit-protocol==1.0.1
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.2.0
nest-asyncio==1.6.0
numpy==2.2.4
onnxruntime==1.21.0
openai==1.72.0
packaging==24.2
pillow==11.1.0
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sniffio==1.3.1
sounddevice==0.5.1
sympy==1.13.3
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.51.1
types-protobuf==4.25.0.20240417
typing-inspection==0.4.0
typing_extensions==4.13.1
Unidecode==1.3.8
urllib3==2.3.0
watchfiles==1.0.5
websockets==15.0.1
yarl==1.20.0
